REFACTOR commands to all reside in the 'src/Commands' directory
