+++ src/Commands/MigrateDbInProduction.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Sfneal\Cruise\Commands\Database;
+namespace Sfneal\Cruise\Commands;
 
 use Illuminate\Console\Command;
 use Illuminate\Support\Facades\Artisan;
